refactor: Adds correct structure to grapes_web_builder
This commit alters the structure of the grapes_web_builder table. It also adjusts the code base
so that the dropdown options now save show the template name correctly, but other information
about the selected template is also available to the grapes_web_page table in the db.

Interim Commit

Interim Commit