fix: return to page one of resutls when filter applied
Adjust code to ensure that if not on page one of result when a facet filter
is applied, the filtered results dispaly from page one. Also addresses error
where if you are on a result higher than the number of results available
in a certain facet filter and apply the filter it does not work.